requirements.txt
Flask==2.3.3
mysql-connector-python==8.1.0
fuzzywuzzy==0.18.0
python-Levenshtein==0.21.1
Werkzeug==2.3.7
requests==2.31.0